+++ .eslintrc.js
@@ -1,3 +1,3 @@
-const { eslintNativeConfig } = require('@dfhernandez/js-utilities');
+const { eslintNativeConfig } = require('@2600hz/js-utilities');
 
 module.exports = eslintNativeConfig(__dirname);