+++ .release-it.js
@@ -1,3 +1,3 @@
-const { releaseItBaseConfig } = require('@dfhernandez/js-utilities');
+const { releaseItBaseConfig } = require('@2600hz/js-utilities');
 
 module.exports = releaseItBaseConfig;