+++ commitlint.config.js
@@ -1,3 +1,3 @@
-const { commitlintBaseConfig } = require('@dfhernandez/js-utilities');
+const { commitlintBaseConfig } = require('@2600hz/js-utilities');
 
 module.exports = commitlintBaseConfig;