+++ src/components/text/TextInfo.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components/native';
-import { getFontSize } from '@dfhernandez/js-utilities';
+import { getFontSize } from '@2600hz/js-utilities';
 import type { TextInfoProps } from './types';
 
 interface STTheme {