+++ styled.d.ts
@@ -1,5 +1,5 @@
 import {} from 'styled-components';
-import { DefaultThemeProps } from '@dfhernandez/native-theme';
+import { DefaultThemeProps } from '@2600hz/sds-native-theme';
 
 declare module 'styled-components' {
   export interface DefaultTheme extends DefaultThemeProps {}